[Refactor] Extract user credit component
